parser-worker/requirements.txt
fastapi
uvicorn
httpx
google-cloud-storage
supabase